requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
distro==1.9.0
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jplephem==2.22
numpy<2.0
openai==1.60.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests
sgp4==2.23
skyfield==1.49
sniffio==1.3.1
squarify
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
pandas
matplotlib
seaborn
SQLAlchemy
astropy